examples/redis/simple-redis-app/pyproject.toml
[tool.poetry]
name = "simple-redis-queue"
version = "0.1.0"
description = ""
authors = ["Thierry <thierrysantoos123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
llama-index-llms-openai = "^0.1.25"
llama-index-agent-openai = "^0.2.7"
llama-index-embeddings-openai = "^0.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
